Split out definitions module.
